requirements.txt
streamlit==1.46.1
altair==5.5.0
blinker==1.9.0
cachetools==5.5.2
click==8.2.1
numpy==1.26.4
packaging==23.2
pandas==2.3.0
pillow==11.2.1
protobuf==4.25.8
pyarrow==14.0.2  # Use pre-built wheel
pydeck==0.9.1
python-dateutil==2.9.0.post0
requests==2.32.4
rich==14.0.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.1
watchdog==6.0.0
appnope==0.1.4
dagshub==0.5.10
mlflow==2.2.2
fastapi==0.115.14
uvicorn==0.34.3
scikit-learn==1.7.0
joblib==1.5.1
python-dotenv==1.1.1

# aiohappyeyeballs==2.6.1
# aiohttp==3.12.13
# aiohttp-retry==2.9.1
# aiosignal==1.3.2
# alembic==1.16.2
# altair==5.5.0
# amqp==5.3.1
# annotated-types==0.7.0
# antlr4-python3-runtime==4.9.3
# anyio==4.9.0
# appdirs==1.4.4
# appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
# asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
# async-timeout==5.0.1
# asyncssh==2.21.0
# atpublic==6.0.1
# attrs==25.3.0
# backoff==2.2.1
# billiard==4.2.1
# blinker==1.9.0
# boto3==1.38.45
# botocore==1.38.45
# cachetools==5.5.2
# celery==5.5.3
# certifi==2025.6.15
# cffi==1.17.1
# charset-normalizer==3.4.2
# click==8.2.1
# click-didyoumean==0.3.1
# click-plugins==1.1.1.2
# click-repl==0.3.0
# cloudpickle==2.2.1
# colorama==0.4.6
# colorlog==6.9.0
# comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
# configobj==5.0.9
# contourpy==1.3.2
# cryptography==45.0.4
# cycler==0.12.1
# dacite==1.6.0
# dagshub==0.5.10
# dagshub-annotation-converter==0.1.10
# databricks-cli==0.18.0
# databricks-sdk==0.57.0
# dataclasses-json==0.6.7
# debugpy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_6a37he2v_t/croot/debugpy_1736267437603/work
# decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
# dictdiffer==0.9.0
# diskcache==5.6.3
# distro==1.9.0
# docker==6.1.3
# dpath==2.2.0
# dulwich==0.23.1
# dvc==3.60.1
# dvc-data==3.16.10
# dvc-http==2.32.0
# dvc-objects==5.1.1
# dvc-render==1.0.2
# dvc-studio-client==0.21.0
# dvc-task==0.40.2
# entrypoints==0.4
# exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
# executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
# fastapi==0.115.14
# filelock==3.18.0
# Flask==2.3.3
# flatten-dict==0.4.2
# flufl.lock==8.2.0
# fonttools==4.58.4
# frozenlist==1.7.0
# fsspec==2025.5.1
# funcy==2.0
# ghp-import==2.1.0
# gitdb==4.0.12
# GitPython==3.1.44
# google-auth==2.40.3
# gql==3.5.3
# grandalf==0.8
# graphene==3.4.3
# graphql-core==3.2.6
# graphql-relay==3.2.0
# gto==1.7.2
# gunicorn==20.1.0
# h11==0.16.0
# httpcore==1.0.9
# httpx==0.28.1
# hydra-core==1.3.2
# idna==3.10
# importlib-metadata==6.11.0
# ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1719845458456/work
# ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1748711175/work
# iterative-telemetry==0.0.10
# itsdangerous==2.2.0
# jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
# Jinja2==3.1.6
# jmespath==1.0.1
# joblib==1.5.1
# jsonschema==4.24.0
# jsonschema-specifications==2025.4.1
# jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
# jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1748333051527/work
# kiwisolver==1.4.8
# kombu==5.5.4
# lightgbm==4.6.0
# llvmlite==0.44.0
# lxml==6.0.0
# Mako==1.3.10
# Markdown==3.8.2
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# marshmallow==3.26.1
# matplotlib==3.10.3
# matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
# mdurl==0.1.2
# mergedeep==1.3.4
# missingno==0.5.2
# mkdocs==1.6.1
# mkdocs-get-deps==0.2.0
# mlflow==2.2.2
# mlflow-skinny==3.1.1
# multidict==6.5.1
# mypy_extensions==1.1.0
# narwhals==1.44.0
# nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
# networkx==3.4.2
# numba==0.61.2
# numpy==1.26.4
# oauthlib==3.3.1
# omegaconf==2.3.0
# opentelemetry-api==1.34.1
# opentelemetry-sdk==1.34.1
# opentelemetry-semantic-conventions==0.55b1
# optuna==4.4.0
# orjson==3.10.18
# packaging==23.2
# pandas==2.3.0
# parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
# pathspec==0.12.1
# pathvalidate==3.3.1
# pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
# pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
# pillow==11.2.1
# platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
# plotly==6.2.0
# prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
# propcache==0.3.2
# protobuf==4.25.8
# psutil @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_10oa1k8l11/croot/psutil_1736367646006/work
# ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
# pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
# pyarrow==11.0.0
# pyasn1==0.6.1
# pyasn1_modules==0.4.2
# pycparser==2.22
# pydantic==2.11.7
# pydantic_core==2.33.2
# pydeck==0.9.1
# pydot==4.0.1
# pygit2==1.18.0
# Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
# pygtrie==2.5.0
# PyJWT==2.10.1
# pyparsing==3.2.3
# python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
# python-dotenv==1.1.1
# pytz==2022.7.1
# PyYAML==6.0.2
# pyyaml_env_tag==1.1
# pyzmq @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_95lsut8ymz/croot/pyzmq_1734709560733/work
# querystring-parser==1.2.4
# referencing==0.36.2
# requests==2.32.4
# requests-toolbelt==1.0.0
# rich==14.0.0
# rpds-py==0.26.0
# rsa==4.9.1
# ruamel.yaml==0.18.14
# ruamel.yaml.clib==0.2.12
# ruff==0.12.1
# s3transfer==0.13.0
# scikit-learn==1.7.0
# scipy==1.15.3
# scmrepo==3.3.11
# seaborn==0.13.2
# semver==3.0.4
# shap==0.48.0
# shellingham==1.5.4
# shortuuid==1.0.13
# shtab==1.7.2
# six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
# slicer==0.0.8
# smmap==5.0.2
# sniffio==1.3.1
# SQLAlchemy==2.0.41
# sqlparse==0.5.3
# sqltrie==0.11.2
# stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
# starlette==0.46.2
# streamlit==1.46.1
# tabulate==0.9.0
# tenacity==9.1.2
# threadpoolctl==3.6.0
# toml==0.10.2
# tomli==2.2.1
# tomlkit==0.13.3
# tornado @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_63s3osc0_1/croot/tornado_1748956943583/work
# tqdm==4.67.1
# traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
# treelib==1.7.1
# typer==0.16.0
# typing-inspect==0.9.0
# typing-inspection==0.4.1
# typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1748959427/work
# tzdata==2025.2
# urllib3==2.5.0
# uvicorn==0.34.3
# vine==5.1.0
# voluptuous==0.15.2
# watchdog==6.0.0
# wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
# websocket-client==1.8.0
# Werkzeug==3.1.3
# xgboost==3.0.2
# yarl==1.20.1
# zc.lockfile==3.0.post1
# zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
